add response to item_view click
