+++ app/src/main/java/com/example/zombiefit/OnFragmentInteractionListener.java
@@ -2,6 +2,5 @@ package com.example.zombiefit;
 
 public interface OnFragmentInteractionListener {
 
-    void onDetailedFragmentInteraction(String title, String image, String description,
-                                       String youTube, String congrats);
+    void onDetailedFragmentInteraction();
 }
